requirements.txt
streamlit
huggingface_hub
python-dotenv
requests